Refactor ongoing transfer components structure
Moved ongoing transfer-related components into a dedicated "ongoing" folder for better organization. Updated import paths accordingly and adjusted the "View on Block Explorer" link placement for improved UI consistency.
